@@ -6,10 +6,10 @@ using System.Text.Json.Serialization;
 
 namespace DeepDungeonTracker;
 
-public class Floor
+public class Floor(int number)
 {
     [JsonInclude]
-    public int Number { get; private set; }
+    public int Number { get; private set; } = number;
 
     [JsonInclude]
     public TimeSpan Time { get; private set; } = new();
@@ -36,39 +36,39 @@ public class Floor
     public int DreadBeasts { get; private set; }
 
     [JsonIgnore]
-    public Collection<Coffer> Coffers { get; private set; } = new();
+    public Collection<Coffer> Coffers { get; private set; } = [];
 
     [JsonInclude]
     [JsonPropertyName("Coffers")]
-    public Collection<Coffer>? SerializationCoffers { get => this.Coffers?.Count > 0 ? this.Coffers : null; private set => this.Coffers = value ?? new(); }
+    public Collection<Coffer>? SerializationCoffers { get => this.Coffers?.Count > 0 ? this.Coffers : null; private set => this.Coffers = value ?? []; }
 
     [JsonIgnore]
-    public Collection<Enchantment> Enchantments { get; private set; } = new();
+    public Collection<Enchantment> Enchantments { get; private set; } = [];
 
     [JsonInclude]
     [JsonPropertyName("Enchantments")]
-    public Collection<Enchantment>? SerializationEnchantments { get => this.Enchantments?.Count > 0 ? this.Enchantments : null; private set => this.Enchantments = value ?? new(); }
+    public Collection<Enchantment>? SerializationEnchantments { get => this.Enchantments?.Count > 0 ? this.Enchantments : null; private set => this.Enchantments = value ?? []; }
 
     [JsonIgnore]
-    public Collection<Enchantment> EnchantmentsSerenized { get; private set; } = new();
+    public Collection<Enchantment> EnchantmentsSerenized { get; private set; } = [];
 
     [JsonInclude]
     [JsonPropertyName("EnchantmentsSerenized")]
-    public Collection<Enchantment>? SerializationEnchantmentsSerenized { get => this.EnchantmentsSerenized?.Count > 0 ? this.EnchantmentsSerenized : null; private set => this.EnchantmentsSerenized = value ?? new(); }
+    public Collection<Enchantment>? SerializationEnchantmentsSerenized { get => this.EnchantmentsSerenized?.Count > 0 ? this.EnchantmentsSerenized : null; private set => this.EnchantmentsSerenized = value ?? []; }
 
     [JsonIgnore]
-    public Collection<Trap> Traps { get; private set; } = new();
+    public Collection<Trap> Traps { get; private set; } = [];
 
     [JsonInclude]
     [JsonPropertyName("Traps")]
-    public Collection<Trap>? SerializationTraps { get => this.Traps?.Count > 0 ? this.Traps : null; private set => this.Traps = value ?? new(); }
+    public Collection<Trap>? SerializationTraps { get => this.Traps?.Count > 0 ? this.Traps : null; private set => this.Traps = value ?? []; }
 
     [JsonIgnore]
-    public Collection<Pomander> Pomanders { get; private set; } = new();
+    public Collection<Pomander> Pomanders { get; private set; } = [];
 
     [JsonInclude]
     [JsonPropertyName("Pomanders")]
-    public Collection<Pomander>? SerializationPomanders { get => this.Pomanders?.Count > 0 ? this.Pomanders : null; private set => this.Pomanders = value ?? new(); }
+    public Collection<Pomander>? SerializationPomanders { get => this.Pomanders?.Count > 0 ? this.Pomanders : null; private set => this.Pomanders = value ?? []; }
 
     [JsonInclude]
     public int Deaths { get; private set; }
@@ -81,8 +81,6 @@ public class Floor
 
     [JsonInclude]
     public MapData MapData { get; private set; } = new();
-
-    public Floor(int number) => this.Number = number;
 
     public bool IsFirstFloor() => this.Number % 10 == 1;
 
@@ -137,5 +135,5 @@ public class Floor
 
     public int Lurings() => this.Traps.Count(x => x == Trap.Luring);
 
-    public IEnumerable<Enchantment> AdjustedEnchantments() => this.Enchantments.Count > 0 ? this.Enchantments : this.EnchantmentsSerenized.Count > 0 ? this.EnchantmentsSerenized : new();
+    public IEnumerable<Enchantment> AdjustedEnchantments() => this.Enchantments.Count > 0 ? this.Enchantments : this.EnchantmentsSerenized.Count > 0 ? this.EnchantmentsSerenized : [];
 }
