@@ -2,10 +2,8 @@
 
 namespace DeepDungeonTracker;
 
-public sealed class ScoreHeavenOnHigh : Score
+public sealed class ScoreHeavenOnHigh(SaveSlot saveSlot, bool isDutyComplete) : Score(saveSlot, isDutyComplete)
 {
-    public ScoreHeavenOnHigh(SaveSlot saveSlot, bool isDutyComplete) : base(saveSlot, isDutyComplete) { }
-
     protected override int FloorCompletionScoreCalculation()
     {
         var total = 0;
