@@ -1,4 +1,5 @@
 ﻿using Dalamud.Interface.Internal;
+using Dalamud.Interface.ManagedFontAtlas;
 using ImGuiNET;
 using System;
 using System.Globalization;
@@ -18,9 +19,9 @@ public class Render
         ImGui.Image(textureWrap.ImGuiHandle, new(width * finalScale, height * finalScale), new(x1, y1), new(x2, y2), color ?? Color.White);
     }
 
-    public Vector2 DrawText(ImFontPtr imFontPtr, float x, float y, string text, Vector4 color, Alignment align, bool calcTextSize = false)
+    public Vector2 DrawText(IFontHandle fontHandle, float x, float y, string text, Vector4 color, Alignment align, bool calcTextSize = false)
     {
-        ImGui.PushFont(imFontPtr);
+        fontHandle?.Push();
 
         var size = (align != Alignment.Left) ? ImGui.CalcTextSize(text) : Vector2.Zero * this.Scale;
         var alignFactor = (align == Alignment.Center) ? 2.0f : 1.0f;
@@ -34,16 +35,16 @@ public class Render
         ImGui.SetWindowFontScale(this.Scale);
         ImGui.SetCursorPos(new Vector2(x, y));
         ImGui.TextColored(color, text);
-        ImGui.PopFont();
+        fontHandle?.Pop();
 
-        return calcTextSize ? this.GetTextSize(imFontPtr, text) : Vector2.Zero;
+        return calcTextSize ? this.GetTextSize(fontHandle!, text) : Vector2.Zero;
     }
 
-    public Vector2 GetTextSize(ImFontPtr imFontPtr, string text)
+    public Vector2 GetTextSize(IFontHandle fontHandle, string text)
     {
-        ImGui.PushFont(imFontPtr);
+        fontHandle?.Push();
         var size = ImGui.CalcTextSize(text) / this.Scale;
-        ImGui.PopFont();
+        fontHandle?.Pop();
         return size;
     }
 
