@@ -5,13 +5,11 @@ using System.Numerics;
 
 namespace DeepDungeonTracker;
 
-public sealed class TrackerWindow : WindowEx, IDisposable
+public sealed class TrackerWindow(string id, Configuration configuration, Data data) : WindowEx(id, configuration), IDisposable
 {
-    private Data Data { get; }
+    private Data Data { get; } = data;
 
     private Action<float, float, float, SaveSlot, FloorSet, Floor>?[]? FieldCalls { get; set; }
-
-    public TrackerWindow(string id, Configuration configuration, Data data) : base(id, configuration) => this.Data = data;
 
     public void Dispose() { }
 
@@ -94,7 +92,7 @@ public sealed class TrackerWindow : WindowEx, IDisposable
         var offsetX = 76.0f + spacing;
         width = columnX - 27.0f + (Convert.ToInt32(config.IsFloorNumberVisible) + Convert.ToInt32(config.IsSetNumberVisible) + Convert.ToInt32(config.IsTotalNumberVisible)) * offsetX;
 
-        this.FieldCalls = new[] { Kills, Mimics, Mandragoras, Mimicgoras, ShowNPCs ? NPCs : ShowDreadBeasts ? DreadBeasts : null, Coffers, Enchantments, Traps, Deaths, RegenPotions, Potsherds, Lurings, Maps, TimeBonuses };
+        this.FieldCalls = [Kills, Mimics, Mandragoras, Mimicgoras, ShowNPCs ? NPCs : ShowDreadBeasts ? DreadBeasts : null, Coffers, Enchantments, Traps, Deaths, RegenPotions, Potsherds, Lurings, Maps, TimeBonuses];
 
         ui.DrawBackground(width, height, (!config.SolidBackground && this.IsFocused) || config.SolidBackground);
 
